Refactor profile settings to own component
